Moving scala conversions to io.finch.syntax.scala.
