+++ core/src/main/scala/io/finch/syntax/Mapper.scala
@@ -3,18 +3,16 @@ package io.finch.syntax
 import com.twitter.finagle.http.Response
 import com.twitter.util.Future
 import io.finch.{Endpoint, Output}
-import io.finch.internal.ScalaToTwitterConversions._
-import scala.concurrent.ExecutionContext
 import shapeless.HNil
 import shapeless.ops.function.FnToProduct
 
 /**
- * A type class that allows the [[Endpoint]] to be mapped to either `A => B` or `A => Future[B]`.
- * @groupname LowPriorityMapper Low Priority Mapper Conversions
- * @groupprio LowPriorityMapper 0
- * @groupname HighPriorityMapper High priority mapper conversions
- * @groupprio HighPriorityMapper 1
- */
+  * A type class that allows the [[Endpoint]] to be mapped to either `A => B` or `A => Future[B]`.
+  * @groupname LowPriorityMapper Low Priority Mapper Conversions
+  * @groupprio LowPriorityMapper 0
+  * @groupname HighPriorityMapper High priority mapper conversions
+  * @groupprio HighPriorityMapper 1
+  */
 trait Mapper[A] {
   type Out
 
@@ -36,8 +34,8 @@ private[finch] trait LowPriorityMapperConversions {
     instance(_.mapOutput(f))
 
   /**
-   * @group LowPriorityMapper
-   */
+    * @group LowPriorityMapper
+    */
   implicit def mapperFromResponseFunction[A](f: A => Response): Mapper.Aux[A, Response] =
     instance(_.mapOutput(f.andThen(r => Output.payload(r, r.status))))
 
@@ -48,24 +46,11 @@ private[finch] trait LowPriorityMapperConversions {
     instance(_.mapOutputAsync(f))
 
   /**
-   * @group LowPriorityMapper
-   */
+    * @group LowPriorityMapper
+    */
   implicit def mapperFromFutureResponseFunction[A](f: A => Future[Response]): Mapper.Aux[A, Response] =
     instance(_.mapOutputAsync(f.andThen(fr => fr.map(r => Output.payload(r, r.status)))))
 
-  /**
-    * @group LowPriorityMapper
-    */
-  implicit def mapperFromFutureOutputFunction[A, B](f: A => scala.concurrent.Future[Output[B]])
-    (implicit ec: ExecutionContext): Mapper.Aux[A, B] =
-    instance(_.mapOutputAsync {x => f(x).asTwitterFuture} )
-
-  /**
-    * @group LowPriorityMapper
-    */
-  implicit def mapperFromFutureResponseFunction[A](f: A => scala.concurrent.Future[Response])
-    (implicit ec: ExecutionContext) : Mapper.Aux[A, Response] =
-    instance(_.mapOutputAsync(f.andThen(fr => fr.map(r => Output.payload(r, r.status)).asTwitterFuture)))
 }
 
 private[finch] trait HighPriorityMapperConversions extends LowPriorityMapperConversions {
@@ -74,31 +59,31 @@ private[finch] trait HighPriorityMapperConversions extends LowPriorityMapperConv
     * @group HighPriorityMapper
     */
   implicit def mapperFromOutputHFunction[A, B, F, OB](f: F)(implicit
-    ftp: FnToProduct.Aux[F, A => OB],
-    ev: OB <:< Output[B]
+                                                            ftp: FnToProduct.Aux[F, A => OB],
+                                                            ev: OB <:< Output[B]
   ): Mapper.Aux[A, B] = instance(_.mapOutput(value => ev(ftp(f)(value))))
 
 
   /**
-   * @group HighPriorityMapper
-   */
+    * @group HighPriorityMapper
+    */
   implicit def mapperFromResponseHFunction[A, F, R](f: F)(implicit
-    ftp: FnToProduct.Aux[F, A => R],
-    ev: R <:< Response
+                                                          ftp: FnToProduct.Aux[F, A => R],
+                                                          ev: R <:< Response
   ): Mapper.Aux[A, Response] = instance(_.mapOutput { value =>
     val r = ev(ftp(f)(value))
     Output.payload(r, r.status)
   })
 
   /**
-   * @group HighPriorityMapper
-   */
+    * @group HighPriorityMapper
+    */
   implicit def mapperFromOutputValue[A](o: => Output[A]): Mapper.Aux[HNil, A] =
     instance(_.mapOutput(_ => o))
 
   /**
-   * @group HighPriorityMapper
-   */
+    * @group HighPriorityMapper
+    */
   implicit def mapperFromResponseValue(r: => Response): Mapper.Aux[HNil, Response] =
     instance(_.mapOutput(_ => Output.payload(r, r.status)))
 
@@ -109,56 +94,25 @@ private[finch] trait HighPriorityMapperConversions extends LowPriorityMapperConv
     instance(_.mapOutputAsync(_ => o))
 
   /**
-   * @group HighPriorityMapper
-   */
+    * @group HighPriorityMapper
+    */
   implicit def mapperFromFutureResponseValue(fr: => Future[Response]): Mapper.Aux[HNil, Response] =
     instance(_.mapOutputAsync(_ => fr.map(r => Output.payload(r, r.status))))
-
-  /**
-    * @group HighPriorityMapper
-    */
-  implicit def mapperFromScalaFutureOutputValue[A]
-    (o: => scala.concurrent.Future[Output[A]])
-    (implicit ec: ExecutionContext)
-    : Mapper.Aux[HNil, A] =
-    instance(_.mapOutputAsync(_ => o.asTwitterFuture))
-
-  /**
-    * @group HighPriorityMapper
-    */
-  implicit def mapperFromScalaFutureResponseValue
-    (fr: => scala.concurrent.Future[Response])
-    (implicit ec: ExecutionContext)
-    : Mapper.Aux[HNil, Response] =
-    instance(_.mapOutputAsync(_ => fr.map(r => Output.payload(r, r.status)).asTwitterFuture))
 }
 
-object Mapper extends HighPriorityMapperConversions {
+private[finch] trait TwitterMapper extends HighPriorityMapperConversions {
   implicit def mapperFromFutureOutputHFunction[A, B, F, FOB](f: F)(implicit
-    ftp: FnToProduct.Aux[F, A => FOB],
-    ev: FOB <:< Future[Output[B]]
+                                                                   ftp: FnToProduct.Aux[F, A => FOB],
+                                                                   ev: FOB <:< Future[Output[B]]
   ): Mapper.Aux[A, B] = instance(_.mapOutputAsync(value => ev(ftp(f)(value))))
 
   implicit def mapperFromFutureResponseHFunction[A, F, FR](f: F)(implicit
-    ftp: FnToProduct.Aux[F, A => FR],
-    ev: FR <:< Future[Response]
+                                                                 ftp: FnToProduct.Aux[F, A => FR],
+                                                                 ev: FR <:< Future[Response]
   ): Mapper.Aux[A, Response] = instance(_.mapOutputAsync { value =>
     val fr = ev(ftp(f)(value))
     fr.map(r => Output.payload(r, r.status))
   })
-
-  implicit def mapperFromScalaFutureOutputHFunction[A, B, F, FOB](f: F)(implicit
-                                                                   ftp: FnToProduct.Aux[F, A => FOB],
-                                                                   ev: FOB <:< scala.concurrent.Future[Output[B]],
-                                                                        ec: ExecutionContext):
-  Mapper.Aux[A, B] = instance(_.mapOutputAsync(value => ev(ftp(f)(value)).asTwitterFuture))
-
-
-  implicit def mapperFromScalaFutureResponseHFunction[A, F, FR](f: F)
-  (implicit ftp: FnToProduct.Aux[F, A => FR],
-             ev: FR <:< scala.concurrent.Future[Response],
-               ec: ExecutionContext): Mapper.Aux[A, Response] = instance(_.mapOutputAsync { value =>
-    val fr = ev(ftp(f)(value)).asTwitterFuture
-    fr.map(r => Output.payload(r, r.status))
-  })
 }
+
+object Mapper extends TwitterMapper
