+++ core/src/main/scala/io/finch/syntax/scala/package.scala
@@ -1,0 +1,6 @@
+package io.finch.syntax
+
+/**
+  * @author Gabriel Francisco <gabfssilva@gmail.com>
+  */
+package object scala extends ScalaMapper
